moving layouts to utils/layout_utils
